Fix pricing
