+++ apps/studio/src/routes/editor/TopBar/Subscription/PricingCard.tsx
