+++ apps/studio/src/routes/editor/index.tsx
@@ -3,6 +3,7 @@ import EditPanel from './EditPanel';
 import LayersPanel from './LayersPanel';
 import Toolbar from './Toolbar';
 import EditorTopBar from './TopBar';
+import PricingPage from './TopBar/Subscription/PricingPage';
 import WebviewArea from './WebviewArea';
 
 function ProjectEditor() {
@@ -29,6 +30,7 @@ function ProjectEditor() {
                     <EditorTopBar />
                 </div>
             </div>
+            <PricingPage />
         </>
     );
 }
